// src/components/units/boardComment/write/BoardCommentWrite.types.ts
import { ChangeEvent } from "react"

export interface IBoardCommentWriteUIProps {

    onChangeWriter: (event: ChangeEvent<HTMLInputElement>) => void
    onChangePassword: (event: ChangeEvent<HTMLInputElement>) => void
    onChangeContents: (event: ChangeEvent<HTMLTextAreaElement>) => void
    onClickWrite: () => void

    contents: string

}